backend/requirements.txt
# ============================================
# CP'S Enterprise POS - Backend Dependencies
# ============================================

# FastAPI & Web Framework
fastapi==0.115.0
uvicorn[standard]==0.34.0
python-multipart==0.0.22

# Database
sqlalchemy==2.0.37
alembic==1.14.1
asyncpg==0.30.0
psycopg2-binary==2.9.10

# Redis
celery==5.4.0
redis==5.2.1

# Authentication & Security
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
cryptography==46.0.5

# Validation & Serialization
pydantic==2.10.5
pydantic-settings==2.7.1
email-validator==2.2.0

# HTTP Client
httpx==0.28.1
aiohttp==3.13.3

# Utilities
python-dateutil==2.9.0
pytz==2024.2
orjson==3.11.5
structlog==25.1.0

# Background Tasks
celery==5.4.0
flower==2.0.1

# Monitoring & Observability
prometheus-client==0.21.1
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-instrumentation-fastapi==0.50b0
sentry-sdk[fastapi]==2.20.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.2
pytest-cov==6.0.0
factory-boy==3.3.1
faker==35.0.0
httpx==0.28.1

# Development
black==24.10.0
isort==5.13.2
flake8==7.1.1
mypy==1.14.1
pre-commit==4.1.0

# Documentation
mkdocs==1.6.1
mkdocs-material==9.5.50

# Optional: Payment Gateways
stripe==11.5.0

# Optional: Email
aiosmtplib==4.0.0
jinja2==3.1.5

# Optional: File Storage
boto3==1.36.6
minio==7.2.15
